requirements.txt
groq
genanki
chromadb
sentence-transformers
pdfplumber
PyMuPDF
python-dotenv
Pillow
rank-bm25
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
sse-starlette>=2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=8.0
